requirements.txt
discord.py
requests
pytz
babel
Pillow
matplotlib
